+++ extension/test/react/fixtures/index.ts
@@ -4,4 +4,3 @@ export { createDashboardPayload } from './dashboardPayload';
 export { createExerciseDetailPayload } from './exerciseDetailPayload';
 export { createGitCredentialsPayload } from './gitCredentialsPayload';
 export { createIrisInitPayload } from './irisInitPayload';
-export { createRecommendedExtensionsPayload } from './recommendedExtensionsPayload';